Fix typos in 2022-07-20-Briefless-Barristers-Lawyerless-Clients.md
I looked closely at the PDF of the typewritten pages, and am fairly
certain that these changes are accurate (but please check my work!)

I avoided taking the liberty to "fix" some punctuation from the
original, such as changing 4 to 6 periods into an elipsis, or correcting
the spelling of "instalment" (that now-rare spelling seems valid
according to some sources).

I actually can't fully parse the grammar of the sentence beginning
"At the same time it sees a large number of citizens...". I think both
"sees" and "seems" work equally well, but "sees" is what appears
in the document.